#IOCP EchoServer Test 문제 없음  2시간째..  추후 계속 테스트 하기

+++ _3Course/EchoServer_ContentsThread/CTestServer.h
@@ -7,7 +7,7 @@ class CTestServer : public CLanServer
     virtual ~CTestServer();
     virtual BOOL Start(const wchar_t *bindAddress, short port, int ZeroCopy, int WorkerCreateCnt, int maxConcurrency, int useNagle, int maxSessions);
 
-    virtual double OnRecv(ull SessionID, CMessage *msg) override;
+    virtual float OnRecv(ull SessionID, CMessage *msg) override;
     virtual bool OnAccept(ull SessionID) override;
 
     void EchoProcedure(ull sessionID, CMessage *message);
